Move controlled field groups
